3.1.5-10 created interface IStackMethod for StackArray and StackList class.
